refactor(comparison): extract setup state
